[IMP] mail, *: add a wizard email configurator * google_gmail, microsoft_outlook, mass_mailing
Purpose
=======
In order to ease the configuration of the email servers, create a modal
that will automatically create the configuration.

This modal will automatically configure the "from_filter", and will
configure the mail.alias.domain of the selected company if we want to
use it as the default mail server.

Task-3061856