[IMP] sale: improve payment link
Expirience of paying through payment link was low, client had no info
what they are exactly paying for other that SO refernce and amount,
no information if current payment confirms the quotation or anything.
Now with link, client has access to SO and they can choose to either
pay amount defined in link or amount that would confirm the order.

task-3087715