[IMP] survey: add language selector
We add a language selector in the first page of the survey enabling users to
select their preferred survey language. This selector is removed once the
survey begins.

We only display the language selector if more than one language is installed
and if the survey title is translated or if at least one question/page has a
translated title.

Technical notes:
- we use the post method to change language to avoid any (future) risk of
infinite direction loop, although the get method was working fine.
- we can't use the language selector of portal as survey doesn't depend on
portal.
- to avoid adding new tours, we test the use-cases in the test_survey tours:
with one language installed, with 2 languages installed but the survey not
translated and with 2 languages installed and the survey translated.
- to avoid adding a new bridge module (website_survey), we add the language
to the default website if the reference website.default_website exists.
- ideally, we should enforce running the tour with website module enabled and
without it as it behaves slightly differently: in the first case, the language
available depends on the default website language and in the other case, on the
installed language. Although the test should work in both case, only one will
be tested by the runbot.

Task-3884012