[IMP] website: add breadcrumb feature to website page
This commit introduces a breadcrumb feature to website pages, enhancing
navigation and user experience. Key features include:

- Breadcrumb Option in Site Properties: Users can enable the breadcrumb
feature from the site properties form view and select a desired parent
page to assign as a breadcrumb to the current website page.

- Customization Options: Users can adjust the background color of the
breadcrumb and modify visibility settings, such as displaying the
breadcrumb over the content or hiding it. Along with that we have also
added two other option to adjust background color and text color for
over the content option.

task-3791063

Co-authored-by: Pratik Raval <prra@odoo.com>