[IMP] core: odoo.domains
New domains.py module to perform domain manipulation.
This will replace the `odoo.osv.expression` module. The domain is now
represented as an immutable abstract syntax tree; it is always normalized
and allows to extend it more easily for things like adding optimizations.

Today domains are manipulated as lists. This adds overhead when trying to
understand the algorithms and you need to normalize manually. You may get
undesired behaviours when you forget to normalize the domain or when you
mutate the list. Concatenating two non-normalized lists is not the same
as the and operator and if a field's search method mutates the domain
that comes from a field's domain, it mutates it for the running instance.

The newer API is fully backwards-compatible and looks like:

    domain = Domain(some_domain_from_user)
    domain &= Domain('id', '>', 5)  # add a condition with an AND
    list(domain)  # get the domain in the old-format

task-3864512