[IMP] survey: add conditional end messages
This commit adds the ability to configure survey completion messages
based on participants' responses.

The functionality is based on a tag system to be added to suggested
answers. Each tag belongs to a survey and has a specific message. A
suggested answer can have several tags.

The completion message displayed will correspond to the tag message most
frequently associated with the user's responses.

In the event of a tie between two or more tags, the 1st tag (in
ascending sequence order) is considered the most frequent tag.

task-3339514