[FIX] web: Fix the issue of Navbar flickering
Before this commit:
Users were facing an issue of flickering when trying to configure a
newly created website. The reason behind this was that the target was
decided after the web client was mounted. When the action target is
evaluated (i.e., new, current, or fullscreen), the navbar was hidden
based on the target and state of the webclient was getting updated using
ACTION_MANAGER:UI-UPDATED event.

What if we already have current_action in browser.sessionStorage, in that
case we should consider current_action.target first.

In this commit, the state of the web client is now set based on the
current_action, if it is not available then we set fullscreen - false.

task-3050040

Co-authored-by: Vivek Panchal <panv@odoo.com>