[FIX] web_editor: resulting wrong cell selection in firefox
Behaviour before commit:

In firefox, if there are table-cells selected through mouse,
aplying background color results in wrong cell selection.
As result wrong cells get colored. This happens because in
firefox selection in tables works with multiple ranges and
it spans <tr> as anchorNode instead of <td>. Due to this,
it results in wrong table selection because of method
historyResetLatestComputedSelection.

Behaviour after commit:

Now it is  possible to apply background color on selected
cells with proper selection. This commit aims to make sure
that selection is not broken before applying color on cells.

task-3433784