[IMP] portal,_*: allow multiple addresses for portal user
Before this PR, customers were unable to manage their addresses autonomously
This could be frustrating as they had to request changes, which might be slow
and would add extra work for sales personnel or customer support.

In this PR, users can add multiple addresses, delete them, and set default
addresses. Additionally, if the user type is a company, the company name,
VAT number and country fields are editable, provided the VAT number is not set.
this allows users to update their information directly.

task-3628329

Co-authored-by: pash-odoo <pash@odoo.com>