[IMP] {website_}event{_*}: pimp up the appointment front-end
*: booth, exhibitor, meet, sale, track

1. Event Form View:

Move some extra fields into into the newly named 'Notes & Documents' page

2. Event Website:

Only display timezone on the website if the visitor is from a different
timezone. Note that the comparison is simply done based on the timezone ID,
which means that Europe/Paris != Europe/Brussels, even if the time is the
same in both.

3. New event_url field:

In order to give more customization to users handling events, add a new
field that is meant, for online events only (no address set), to allow
redirecting to an external website page. If the field is set, the mail
templates (and website confirmation page when website is installed) will
use that url when redirecting to the event.

The 'View Tickets' will be replaced with 'View Event' button for online
events overall. As the tickets will not be used for online events, this
makes the event web page available instead. Again, the links will use
event_url if set, then the website_url if the event is published, else
the button will not appear / the hyperlink will not be set in the mail
templates.

UPG PR: https://github.com/odoo/upgrade/pull/6374
Task-3805889