[FIX] base, tools: fix the behaviour of inline translated elements
Steps to reproduce:

- Go to a website page (in "Edit" mode) > Add a "Title" snippet.

- Select the whole "title" > Set a highlight effect on it.

- With the same text selected, transform it to a link and save.

- Try to translate the title into another language in the editor > You
cannot update the text or remove it.

The code from [1] introduced a feature that allows forcing some specific
elements to be "translated inline" using the `o_translate_inline` class.

It was used mainly to handle specific cases where elements DOM is
handled in JS in a way that breaks the "Translate" editor (e.g., text
highlights).

By forcing the highlight `<span/>`s to be inline translated, the fix
from [1] only handled situations where a highlight effect has a non-
"inline-translated" element (e.g., a link) amongst its text content.

E.g. This DOM structure:

```
<span class="o_text_highlight o_text_highlight_wavy o_translate_inline">
    Go to the <a href="/contactus">Contact Us</a> page
</span>
```

Which caused the translation `<span/>` to be set inside the highlight
structure [A].

Now, as explained in the steps above, there are some cases like with
this DOM:

```
<span class="o_text_highlight o_text_highlight_wavy o_translate_inline">
    <a href="/contactus">Contact Us</a>
</span>
```

That will make the translation code check the inner non-"inline
translated" children even within a `o_translate_inline` parent (see:
`translate_xml_node()` > `hastext()` | `process()`), leading to the
same issue as [A].

The goal of this commit is to prevent any similar situation causing the
translation spans to be added inside elements forced to be translated
as a whole, by automatically considering all elements inside a
`o_translate_inline` parent as inline translated too.

This commit also adds a test for the explained behavior.

[1]: https://github.com/odoo/odoo/commit/e9659e55356e5e4de63f2324eb88a7b6f07cd835

opw-4243639

linked to: opw-3980975
linked to: opw-4089482
linked to: opw-4061566