[IMP] sale_purchase_project: add reinvoiced SO stat button in PO form
To be consistent with the reinvoicing in expenses, we introduce a new stat button
which links the reinvoiced sales orders from the vendor bills of the purchase order.
This stat button appears in the form view of purchase orders.
The idea is to be more transparent to the user by showing the smartbutton so that they would
understand that the line was added there automatically.
Moreover, when multiples lines have been reinvoiced, it may be difficult to find which sale orders
have been impacted.

task-4196812