[FIX] website: fix `s_empowerment` snippet width
This commit aims to fix a width issue within the `s_empowerment` snippet.

Prior to this commit, the `container` of the snippet was set to `fluid`.
While this was fine at a `1440px` width device, it looked off at a larger
one.

This commit changes that class from `container-fluid` to `container` to
ensure the best visual result at all size.

task-4211212