[FIX] html_editor: prevent (up/down) arrow keys from skipping icons
Steps to reproduce:

- Create three paragraphs.
- Insert icons into the second paragraph.
- Place the cursor in the first or third paragraph and press the up/down
arrow key, it skips the second paragraph.

- Set up following HTML:
```html
<p><br></p>
<p><br><span class="fa fa-glass"></span>[]</p>
```
- Press arrow up key.
- Cursor will move to the first paragraph.

Before this commit:

- Pressing the up/down arrow keys would skip a sibling block
containing only icons.
- Pressing the up/down arrow keys would skip over icons in the same
block if the cursor is on a `<br>`.
- For `ArrowUp`, the cursor would move to the first block when on an
icon in the second block with `<br>`, and vice versa for `ArrowDown`.

After this commit:

- Move cursor after icon when `<br>` follows icons within same block.
- For `ArrowUp`, move the cursor to the `<br>` instead of the previous
block when the cursor is on an icon in the second block,
and vice versa for `ArrowDown`.
- For `ArrowDown`, move the cursor after the last icon when in
the last line of the first block.
- For `ArrowUp`, move the cursor to the last icon of the first block
when on the first line of the second block.

task-3045033