[REF] website, html_editor: fuzzy page search for linkpopover url input
In this PR, we've updated the linkpopover component of the html_editor
to add a fuzzy page search feature to the URL input. Before 18.0, we
created another OWL app to load the page anchors. Here we do it by using
autocomplete component, users can now see and select from a dropdown of
relevant Odoo page anchors as they type in the URL input field.

To achieve this, we modified the popover template to include an
autocomplete field specifically designed for the link popover. In the
AutoCompleteInLinkpopover component, two new props were introduced:

1. inputClass: Allows customization of the input field's styling.
2. updateValue: Enables updating of the link popover's URL state when an
option is selected from the autocomplete dropdown.

We also remove the redundant linkpreview reload when clicking on apply,
cause the loadAsyncLinkPreview is already called at onMounted.

task-4222657