[IMP] setup: autoinstall script
A common requests targets to have a script able to setup an odoo env
easily. This is a first poc of such script.

The users of such a script could be a employee setuping a new computer,
a developer willing to test odoo devlopment, a sysadmin setting up a new
server, ... This is why the script should be customisable
but still easy to use for a basic install. The scripts should be able to
be run multiple time, to add some new components.

In the most minimal version, the script will only install the strictly
necessary dependencies to start odoo-bin without an error.

The most complete part of this script will give a result close to the
docker images used on runbot, with all optionnal dependencies, tools,
checker.

This script also uses what is considered as a good practices in production
using debian packages, and not pip packages.

Note: No cleanup or uninstall is managed by the script.

Summary of some goals:
- be usuable in automated scripts, higly customisable using command line
- standalone, downloading this file should be enough
- usable to setup a dev environment
- usable to partially setup a prod environment
- prepare support for multiverse