[IMP] l10n_ee: rework tax report
The tax report (KMD report) and annex to the tax report (KMD INF report)
for Estonia needed to be reworked to correct errors and fix rounding
issues.

In this commit:
- the 'balance_from_tags' label is added to the tax report, allowing a
better audit of the report lines.
- the tax tags are edited to differentiate between tax and base tax lines.
- the loss and profit account fields for the rounding differences are
added to res_company
- default values to these fields are added in the Account Chart Template.

These changes are necessary to improve auditing of the tax report lines and
correct closing entries.

task-3997203

closes odoo/odoo#190448

Related: odoo/enterprise#75530
Related: odoo/upgrade#6924
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>