[FIX] calendar: default privacy not working for internal users
Before this commit, the calendar default privacy setting was not working
for internal users beacause they couldn't see the setting value of other
users if they weren't administrators.

This commit fixes the issue by making a sudo update on the ResUsersSettings
object when computing the calendar default privacy setting of it (and in the
inverse method as well). Sudo was needed to allow the business customization
and bypass the security that the record rule dictates on users not being able
to see the other user's settings (unless they're administrators).

task-4260794

closes odoo/odoo#188744

X-original-commit: ee6f5a6e522df472e3a74eeb102fd242c5e1901b
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>