[IMP] product,uom: remove uom category and merge packaging into uom
This commit introduces major changes to units of measurement and product
packaging features:

*** Removing UoM categories:

Previous to this commit, units were classified into categories such as
length, weight, etc. If the user defined a base unit for a product, any
movement on that product (stock move, order line, etc.) can be only done
in a unit in the same category of the base unit of the product. This
commit removes unit categories and the previously mentioned enforcement.
This is useful for some business cases that need more flexibility (for
example: buying from vendors in pieces and selling in kilograms, etc.).
To allow for some structure, a new field is added to `uom.uom` model to
define a reference unit for any unit. This will be explained further in
a later section.

*** Merging `product.packaging` model into `uom.uom`:

Previous to this commit, users could activate packagings setting to be
able to define packagings for specific product. This caused some
confusion as there is no intrinsic difference between a packaging and a
unit. Another source of confusion is that `stock.move`,
`sale.order.line`, `purchase.order.line`, etc. had two different fields
for the unit and the packaging. They also had two different fields for
the quantity, one based on the unit of the line and the other based on
the packaging. This was confusing for some users and there was also some
technical complexity to mainatain the consistency between those fields.
This commit removes `product.packaging` model completely and all of its
related fields. If the user wants to define some packaging, they just
need to create a new unit and define its quantity (and optionally its
package type). Then, the user can make this packaging usable on some
product by adding it to the new many2many field "Packagings" `uom_ids`
on `product.template`. Users can also define one (or more) barcode for
the packaging of a specific product.

*** Cleaning `uom.uom` model:

This commit removes some of the unnecessary fields on `uom.uom` model
after removing UoM categories. Category, type and color are no longer
needed. Ratio and inverse factor are also removed. Factors now are
defined a bit differently. Now, each unit can (optionally) have a
reference unit (`relative_uom_id`) on which the user-defined
factor (`relative_factor`) is based. Since the relative unit can have
another relative unit on its own, a hierarchy of units can be made.
Therefore, the absolute factor (`factor`) needs to be computed behind
the scenes to be used in quantity and conversion calculations. The
rounding field is also removed from `uom.uom` model because this
introduced some inconsistency with the decimal accuracy record "Product
Unit of Measure" defined on the system level. After this commit, all
conversions related to UoMs are rounded based on that value.

*** Cleaning UoM data:

This commit removes some non-commonly used units that were automatically
installed with `uom` module. Some imperial units are now inactive by
default and activated on installing the US localization. The units that
are totally removed are: km, mi, cm, oz, gal, qt, fl oz, cubic foot and
cubic inch. The units that are deactivated: in, ft, sq ft and lb.

Enterprise: odoo/enterprise#72206
Upgrade: odoo/upgrade#6629

Task-4252043