[IMP] website: enhance `s_blockquote` snippet
This commit enhances the default `s_blockquote` by adding two improvements
to its behavior and design.

1. As we did for `s_card` in commit
e0fc83760f991fa2fea39763bb2709d9e0182316, this commit introduces the
ability to bind `s_blockquote` options to its parent column.

2. The user can now toggle a snippet option to display a `s_rating` inside
the `s_blockquote`, to achieve a layout oriented on
customer reviews.

These changes allow to ease the edition of the snippet (avoid duplicated
options in the case of [1]), and unlock new layouts for [2], which will
simplify the flow for the users.

task-4264114
part of task-4077427

Co-authored-by: Brieuc-brd <brd@odoo.com>