[IMP] product, stock: merge unpriced product labels, add base prices
Priced and unpriced product label formats were merged to avoid
duplicates. The option to print the price is now a separate option in
the view. The pricelist field is hidden from the view if the price
isn't set to be printed. Additionally, it is now possible to print the
price on every product label format, as well as the base unit price,
if enabled and set.

Additionally, the ZPL product label template was made to be generic
enough to work without `stock` and was moved to `product`.

Task ID: 4277792