[IMP] base: stop logging INFO access errors
There is no need to log an INFO every time we create an AccessError.
The exception will be raised and the caller is responsible of either
letting it propagate or catching and handling it: by replacing the
exception, logging, etc.
We were always logging a single INFO line when creating the exception
without a stack trace, even when it is caught and handled.

In the WSGI app handler, we add the stacktrace of the AccessError in
order to have more information for investigations.