[REM] core: remove current_thread().testing
For testing indicator, we have: `odoo.modules.module.current_test` flag.
We should not rely anymore on setting the value on the current thread as
it duplicates the existing indicator; so this method is removed.

When loading demo data, we should not rely on the `testing` flag.
Actually, we may not set it globally because of concurrency issues.
Therefore, demo data is not loaded anymore with a testing context.

For sending e-mails, we deactive them during testing unless a mocked
SMTP server is set up to catch them.
Sending the e-mail invalidates the caches, since we are not doing it
anymore, some additional fixes are necessary: flush a missing field in
mail_activity_mixin, invalidate fields during testing.

task-4270485