[FIX] sale_{stock,stock_margin}: correct margin when SOL added from delivery
**Current behavior:**
Having a sale order which is confirmed and has a pending
delivery, adding a new delivery line which will in turn add a
new SOL to the sale order will not properly compute the margin
for this new line.

**Expected behavior:**
Accurate margin result.

**Steps to reproduce:**
1. Create a stored product with some price and change its
invoicing policy to 'delivery'

2. Create a new sale order, add 1 of the created product,
then confirm it

3. Open the delivery, add another line to it for the same
created product, fill in the quantities for each move, then
validate the picking

4. Check the sale order and see that the new line is not
correctly impacting the order totals and when the `margin` field
is un-hidden from the field/column selector on the notebook that
it is (wrongly) 0.

**Cause of the issue:**
Order lines added in this manner have `product_uom_qty == 0`
which will lead to the inaccuracy.

**Fix:**
When a SOL is created in this manner (from the delivery picking)
let `product_uom_qty = move.quantity` (where `move` is the stock
move added on the delivery picking) if the product is invoiced
on delivery.

opw-4210056