[FIX] portal: fix customer information form if no country set
Having no country on a portal user would cause attempting to edit
personal information through the portal to have an internal server
error.

Added a check to ensure that the partner has a country_id before passing
it to the frontend as a str. Then checking if the partner_id is false instead of
having a string 'false' being passed.

opw-4309159