[FIX] website: fixing carousel transition issue
Step to reproduce:
1. Drag and Drop a Carousel Snippet.
2. Change the Transition from Slide to Fade.
3. Remove all the background images and color.
4. The transition malfunctions

Issue:

The carousel transition CSS currently has an ease-in and ease-out
duration of "0.6s", causing the old text to be removed and the new text
to be inserted simultaneously. This setup results in a visual
malfunction in the transition.

Solution:

To solve this issue we are increasing ease-in time to "3s" and redusing
ease-out to "0.2s".