[IMP] website, *: remove company-specific social media
This commit removes company-specific social media from the website
module. The decision to remove these features was made to avoid
confusion for users regarding which social media links are synchronized
and which are not.

Key Changes:

* Resolved outdated social media routes in company data by mapping
them to real values.
* Updated the FontAwesome fa-youtube icon to fa-youtube-play for better
compatibility.
* Ensured social media icons dynamically update when the associated
link changes.
* Fixed an issue where the social media icon was not visible in the
social snippet section.
* All social media links to have on/off toggler button along with
fa-minus to remove it as well.
* Replace the fa-home icon wiht fa-pencil for footer as well.
* Added test cases to validate social media icon changes, ensuring
functionality and maintainability.

task-4297366