[IMP] website: add parallax zoom effects
For better customization I added the Zoom In and Zoom Out parallax
 effect on background images.

The parallax effect on background images before had only "None","Fixed",
"Bottom to Top" and "Top to Bottom".

Now the parallax effect has "None", "Fixed", "Parallax to Top",
"Parallax to Bottom", "Zoom In" and ''Zoom Out".

I also renamed:
	- "Parallax" to "Scroll Effect",
	- "Bottom to Top" to "Parallax to Top",
	- "Top to Bottom" to "Parallax to Bottom";
to have a better naming for the options.

task-4317372