[IMP] html_editor: document tab in media dialog
This commit changes the rendering of the uploaded document via the media
dialog to the same one from the /file command: an embedded component
with editable file name.

Is also makes the media dialog extendable, allowing the additon of
custom tabs.

The old rendering of the uploaded document (a simple link) is kept as an
alternative to using embedded components.

task-3522395