[FIX] hr_timesheet: resolve access error on opening timesheet by portal user
Steps to reproduce:
- Install project and timesheet app.
- Create a project.
- Create the task.
- Add a timesheet to the task.
- Share the project with the portal user with edit rights.
- Login from the portal user on mobile.
- Open project.
- Open task form view.
- Go to the timesheet tab.
- Click the timesheet record.
- An access error related to the analytic account occurred.

Issue:
An access error related to the analytic account occurred.

Cause:
Due to the change in the web_read of the analytic account model. When the
analytic account field loads it sets analytic_plan_id context.  So, when the
portal user opens the timesheet record it opens the analytic line form view and
there is an analytic account field (account_id) that tries to set the context of
analytic_plan_id. Due to the portal user not having access to the analytic
account an access right error occurred.

Solutions:
1) Provide access to the analytic account field.
Just to display the field, it is not the right way to provide access.
It may be dangerous in the future.

2) Hide the field.
As we cannot create a new view in the stable version, we can set the groups on
the account_id field in the analytic line form view. So, the field is not
visible to portal users.

task-3969988