[FIX] account: no tva recomputation on partner change
Description of the issue this commit addresses:

In an invoice's form view before confirmation, it is possible to edit the tax
amounts in the subtotals of the invoice but those amounts are recomputed when
the partner of the invoice is changed. The way it's currently makes it so that
we lose the custom amounts on such events. We don't want that;

Steps to reproduce:

1. Install account
2. Create a new Vendor Bill, no partner, at least one line with tax. Save.
3. Modify the tax's amount in the subtotal section (click the pen). Save.
4. Assign a partner to the invoice.
5. The custom tax amount is reset to it's rate * the base amount

Desired behavior after the commit is merged:

The tax amount doesn't change when the base amount of the tax doesn't change
and the new tax amount is not the original (reverted changes).
The assignations in the diff are actually putting the current value back in the
tax group values dict instead of using the component's new one.

task-4134532