[FIX] web: pdf.js print on Android device preview got a blank page
This commit fixes the blank page when we print some document.
These bug is known for the pdf.js lib[1] but give a won't fix in the
upstream pdf.js libs because is platform specific[2].

So we patched the lib with the fix given inside the thread issues[3] to
avoid the blank page.

Steps to reproduce:

- Open Odoo on the Android mobile App
- Go to "Shop Floor"
- Click on 'WH/MO/00003' > 'Assembly 1' > 'Worksheet'
- Click the print button on pdf.js toolbar
=> blank screen

opw-4190135

Co-authored-by: Jean-François Aubert <ajf@odoo.com>

[1]: https://github.com/mozilla/pdf.js/issues/10630
[2]: https://github.com/mozilla/pdf.js/issues/10630#issuecomment-817657532
[3]: https://github.com/mozilla/pdf.js/issues/10630#issuecomment-855754913