[IMP] html_editor: file card style
The file card resulting from the /file command (or from uploading a file
via the media dialog) has been simplified and its style has been
changed. Now it is rendered as an inline element allowing for grid
display of multiple files.

The download button has been removed from the file card, but an empty
toolbar is still present to allow for extension in the Knowledge module.

task-3522395