[FIX] purchase: fix product packaging changes
Steps to reproduce:

1. Enable the option "Product Packagings" on the sale/purchase configuration page.
2. Create two or three packaging on a product (eg: Small box, Box and Pallet)
3. Create a new purchase order and select a supplier
4. Select a product with at least two packagings
5. Change the packaging (eg Small Box)
6. Change the packaging quantity (eg 1 small box)
7. Change the packaging (eg box)
8. Change the packaging quantity (eg 1 box)

Issue:

The suggested_packaging overwrite the current packaging type, changing
the quantity and the current packaging. This issue confuses customers,
as they do not understand the changes, leading to misunderstandings.
My fix simplifies the process and makes it easier for customers to comprehend.
By simply not changing the packaging if one is already set.

Fix:

Ensure no packaging is set on the purchase line before searching the
suggested_packaging.

opw-4361531