[FIX] purchase_stock,stock_landed_costs: link landed cost on bill
Steps to reproduce

1. Create a product invoiced on ordered quantities
2. Create a PO and a bill for it, post the bill
3. Receipt half the quantity
4. Add a landed cost product on the PO, create a second bill for this
   landed cost product. Create and validate the landed cost
5. Receipt the second half quantity.

Issue

The valuation of the second transfer is too low

Explanation

To get the price unit for "on ordered qty" policy, We compute the ratio
of already receipt value on already invoiced value

https://github.com/odoo/odoo/blob/4f6353ec8a7306fbf63e95e8e02248d72f413aa3/addons/purchase_stock/models/stock_move.py#L84

The issue is the receipt value is increase by the landed costs value
because the receipt valuation layer is linked to the landed cost
valuation layer. But the product invoice line is not linked to the
landed costs invoice line. Resulting in a invoice value under valuated.

This commit will increase the `total_invoice_value` by the same value of
the landed costs value *if* the landed cost has it's own invoice line.
In case the landed cost is created manually and it's value is already
added into the Purchase order line price unit, we don't want to
artificially increase the invoice value.

Task : 4354498