[FIX] mail: defensively add message_id in references
When email provider rewrites message_id of odoo generated emails
we lose trace of original message and/or trace in Odoo. Crafting
references as already containing the message_id ensure to keep
a trace even if message_id is rewritten.

Continuation of odoo/odoo#... but this time for mailing generated
mail records.

Task-3927616