[IMP] mass_mailing_sms, sms, test_mailing_sms: don't display canceled message
For mailing only, in order to not display most message not sent to the
recipient in the chatter, while preparing message to be sent by sms, we only
create not canceled sms (state != 'canceled') and create directly a trace for
further investigation.

The sms sent without a mailing will be handled in another PR using mail
notification instead of traces for handling errors.

Task-3264570