[FIX] website_google_map: world map zooming out when trying to render a partner without a location
The World Map view when you are on the "/partners" page doesn't zoom. This happens when any of the partners
that are getting rendered doesn't have location.

So, we can filtered out those partners that don't have the location, and only render the ones that have.

Steps to Reproduce on Runbot:
1. Integrate the Google Place Map key via settings->Geolocation->Google Place Map
2. Go to "/partners"
3. Make sure to have a world map view toggled on for the page via editor.
4. Make sure one of the partner on the page doesn't have location.
5. Go to the "map" view, and we'll see that the map is not zoomed in.

opw-4319619