[IMP] web_editor: skip media dialog on file upload from linktools
This commit changes the behavior of the file upload button in the
website sidebar's link tools. Clicking on it will skip the media dialog
and directly open the system's file selector, allowing for a fast upload
flow. The button's icon has also been changed to an upload icon to
better convey its action.

task-3522395