[FIX] website_payment: remove unnecessary toaster warning during payment
This commit addresses the issue of unnecessary toaster notifications
that occur when payment processing fails. The root cause of these
notifications is the Promise.reject call within the _initiatePaymentFlow
function in the payment module. To resolve this, we have wrapped the
_initiatePaymentFlow function in a try-catch block. This allows us to
handle errors more gracefully and modify the warning messages
accordingly.

task-4191431