[FIX] survey: create empty answer when token access set
... to allow general print of the survey.

When we set access mode to token, we need
to have existing registration `survey.user_input` to print the survey.

This commit suggest solution similar to the one existing in survey_test,
so if we create test answer to use it.
Reproduce
---
- -i survey
- create new Survey, with "Access Mode": "invited people only"
- click "Print" BUG -> taken to the odoo main page, instead of empty printed survey

Def
---
What I refer to as "General print" of the survey is a request to print survey without answer_token.
This is currently possible to request from the form view of the survey.survey

opw-4292331