[IMP] website_sale: distinguish checkout rental item
The goal is to visually distinguish rented items from purchased items in
the Order Overview of the checkout process.

This commit adds a container above `o_cart_product` to be able to
conditionnally add a title after the last rental_items of the
website_sale_order. Follows the enterprise changes made in the
website_sale_rental.

task-3531721