[IMP] hw_drivers,point_of_sale: certification for scales
In order to comply with the EU Measuring Instruments
Directive, the code for weighing products must be
verified with a checksum to ensure it has not been
modified. This commit refactors both the IoT driver
and the POS code, with the goal of removing dead
code and making the amount of code that needs to
be certified as small as possible.

task-3929839