[IMP] website_sale: hide rental period if no rental item
Currently on the cart if the users delete all rental products, the
rental period stays in the UI.

This is conditionned with qweb variables thus won't be applied until
page refresh. This commit adds a Js condition to hide the rental period
if all rental product are deleted from the cart.

task-3531721