[IMP] sale_purchase{,_stock},stock_dropshipping: dropship addr fixes
If a dropship purchase order was created automatically through a sale
order, the delivery address field must be readonly. If the sale
order's shipping address changes, the associated purchase order's must
also change to reflect this. Similarly, the delivery address for the
lot must be the delivery address rather than the client company's own
address (which can differ with the customer addresses option enabled)
if the delivery is a dropship.

Task ID: 4291198