[ADD] hr_holidays: Correct By Employee report.
When an employee takes a leave that spans more than one month,
for example from 15/10 to 13/12, the report By Employee will
show the start month (October) and the total number of days
of the leave request instead of showing each month spanned by
the request and the number of days of each month.

To fix the issue, a new report hr_leave_employee_report is created
instead of using the data of hr_leave directly in the report.
In such a way, it is possible to customize the data shown in the
reports to show the expected result.

task-4237738