[IMP] website_sale: review cart_lines and combo layout
Currently, the products displayed in the /shop have a bad
responsiveness. On mobile viewports the quantity input goes above the
description and the product title, making it unreadable.

The combo items layout on mobile doesn't work. On some screensize only
the picture of subitems are visible, thus missing important informative
content.

In this commit:
- Review `cart_lines` template
- Review `o_cart_product` layout for improved responsiveness
- Combo sub items now use a list-group, no more image.
- Improved save for later / remove button design
- Reviewed font-size to match condensed layout (14px)
- Better text truncation to ensure attribute readability yet cropping
long descriptions.

task-4387239