[FIX] survey: fix close buttons
- Install survey
- Create a survey “Live Session”
- Add a “Single Line Text Box” and check “Save as user nickname”
- Add a second question (ex.: multiple choice with 1 correct, score: 1)
- Click on “Create Live Session”
- Join the session with another browser
- Answer the question correctly
- On the survey manager, go to the end of the survey (Final leaderboard)
- Click on the "Close" button

Nothing happens while it should close the session and get back to the survey
form.

Actually, the code was closing the session but failed to get back to the survey
as it was using "window.history.back()" and the button "create session"
launches the session in a new tab (with no history). We solve the problem by
redirecting the user to the survey.action_survey_form instead of using
history.back().

Note: unfortunately, we cannot change easily the tour so that it fails with
history.back() as we can't open a new tab in the tour (see comment in the
method action_open_session_manager_mock).

Task-4381603