[FIX] account: Fix various issues with round globally
-------------------------------------------
Fix the "same_tax_base" on tax totals.
-------------------------------------------

Suppose 2 lines:
- 100 with 13%
- 100 with 23%
=> The base are not displayed when printing the PDF because 'same_tax_base' is True.
However, the untaxed amount is 200 and both taxes have a base of 100.

-------------------------------------------
Fix the base amount when there is no tax.
-------------------------------------------

Suppose 2 lines having quantity=12.12, price_unit=12.12
=> The expected base amount is round(12.12 * 12.12 * 2) ~= 293.79
However, the current dispatching of taxes is only done when there is a tax set.